cleanup comments and unused files
