+++ citation-server/package.json
@@ -16,17 +16,14 @@
   "license": "ISC",
   "dependencies": {
     "assign-deep": "^0.4.5",
-    "autoprefixer": "^6.7.7",
     "babel-polyfill": "^6.22.0",
     "citation-backoffice": "^0.2.0",
     "citation-react-builder": "^0.2.0",
     "citation-react-renderer": "^0.2.0",
-    "extract-text-webpack-plugin": "^2.1.0",
     "fs-promise": "^1.0.0",
     "graphql": "^0.7.2",
     "hapi": "^15.2.0",
     "hapi-graphql": "aya-experience/hapi-graphql",
-    "html-webpack-plugin": "^2.28.0",
     "inert": "^4.1.0",
     "lodash": "^4.17.2",
     "meow": "^3.7.0",
@@ -34,7 +31,6 @@
     "node-cron": "^1.1.3",
     "nodegit": "^0.17.0",
     "os": "^0.1.1",
-    "webpack": "^2.4.1",
     "winston": "^2.3.1"
   },
   "devDependencies": {
