+++ citation-server/src/gitupdater/components.js
@@ -4,7 +4,8 @@ import winston from 'winston';
 import cron from 'node-cron';
 
 import conf from '../conf';
-import render, {spawn} from '../rendering';
+import render from '../rendering';
+import spawn from '../utils/spawn';
 import {updater} from './updater';
 
 const logger = winston.loggers.get('GitUpdater');
@@ -17,7 +18,6 @@ export async function updateCallback() {
 	const masterPath = path.resolve(directory, 'master');
 	await spawn(componentsConf['install-command'], masterPath);
 	await spawn(componentsConf['build-command'], masterPath);
-	// await spawn(componentsConf['compile-command'], masterPath);
 	await render();
 }
 
@@ -27,7 +27,13 @@ export async function start() {
 	logger.info(`Components updater starting on ${cronValue}`);
 	const repository = componentsConf.repository;
 	const directory = path.resolve(conf.work.components, '0');
-	const boundUpdater = updater.bind(undefined, 'Components (0)', repository, directory, updateCallback);
+	const boundUpdater = updater.bind(
+		undefined,
+		'Components (0)',
+		repository,
+		directory,
+		updateCallback
+	);
 	await boundUpdater();
 	tasks[0] = cron.schedule(cronValue, boundUpdater, true);
 }
